src/requirements.txt
psycopg2-binary==2.7.5
Django==2.1.5
flake8==3.6.0
pylint==2.2.2
django-extensions==2.1.4
ipython==7.2.0
Pillow==5.3.0
social-auth-app-django==3.1.0
sorl-thumbnail==12.5.0
mod_wsgi==4.6.5
django-debug-toolbar==1.11
redis==3.0.1
werkzeug==0.14.1
django-braces==1.13.0
django-crispy-forms==1.7.2
djangorestframework==3.9.1
django-filter==2.1.0
